Conditionalise the Fedora/EPEL spec Qt usage for EPEL5
One of the (hopefully last) remaining issues with the pcp-gui
merge was uncovered while doing the release builds - the spec
for EPEL5 attempts the pcp-gui build unconditionally.  Turns
out this is a problem, and we weren't building pcp-gui at all
on EPEL5 before - due to the age on the Qt packages there.

But, no reason not to have pcp-doc there ... so now we can do
that.  This change conditionalises the spec so that we can go
forward with almost-identical specs and future release builds
will be uncomplicated once more.