Make the collectl test faster and platform-independent once more
In the leadup to 3.9.4 we added a last minute fix to pmcollectl,
and tested it via an extension to qa/709.  This commit builds on
that extension, removing Linux-specific aspects of those changes.

Additionally, I noticed the test was starting to take a fair bit
of time (~1 minute) and getting longer with each extension which
seemed a bit odd.  Turned out to be relatively simple to get the
test to execute more quickly by reducing the sampling interval.

Further work remains there, once pmcollectl is able to record at
sub-second intervals (RH BZ 1098117).  With this change, we have
carved off 20-30 seconds or so, with that other change we should
get the test down to running in less than 10seconds all up (it's
at around 40sec elapsed time now).

Which all helps when trying to debug/run the test multiple times
in the leadup to a release.  :)