Fix pmlogcheck bad memory access sigsegv on some logs
If an archive contains a metric which initially has an
error recorded for it (no value) but later on good data
arrives, and that metric is a singleton (PM_INDOM_NULL)
then pmlogcheck will fail to handle it correctly.  This
is because no space to hold the previous value for any
fetch is allocated during the initial error-containing
sample, and subsequently an assumption is made that we
did allocate that space.

Metrics with instances were handled correctly all along,
this only affects single-valued metrics.

Test 701 is added to exercise the fix.  An audit of the
existing QA archives found the ace_v2 archive contains
the problematic pattern, and without the fix produces:
701: line 21: Segmentation fault  pmlogcheck src/ace_v2

This resolves Fedora BZ #1077432.