Handle additional modern distros having /var/run on tmpfs
At some point along the way, Fedora and RHEL began using a non
persistent filesystem to back /var/run, and so we should not be
installing files there permanently.  Now handled the same way
we tackled this on other platforms with similar setups - since
the scripts can already create the necessary paths themselves,
on the fly, let 'em do so.

While auditing this, found a couple of cases of needless setting
of a RUNDIR variable - in the pmlogger and pmie rc scripts.

Also while auditing this, found a case of open coded pcp:pcp use
in a chmod, instead of using the $PCP_USER:$PCP_GROUP variables
as all the other scripts do.

This resolves Fedora BZ #656659.