Fix daemon signal handlers to not make use of unsafe functions
It is unsafe to make certain library calls from signal handlers,
and in several daemons we do this for several different reasons.

Firstly, some daemons trap sigsegv and sigbus to attempt to get
a stacktrace in their log files.  This should be deferred to any
system daemons that do this, so we now do that - by default.  In
certain situations those daemons may not be running or platforms
may not have such daemons - thus there are situations where we
may still want to attempt to generate a stackdump in a last-ditch
attempt at diagnostics.  Thus, these remain as a debug option via
DBG_TRACE_DESPERATE for those occassions.

Similarly, most daemons printed diagnostics showing which signal#
they received before exiting, which is also unsafe - again, move
these to desperate-level diagnostics, off-by-default.

A third case was pmie, which rotated log files on receipt of USR1.
This is reworked to be deferred to outside of the signal handler.

The fourth and final case was pmlogger handling of volume switches
via HUP - this has been similarly reworked such that its deferred
to outside of the signal handler.

This resolves Red Hat BZ #847343.