Tweak settings in pmquery application for neater output
Two changes here - make the default window size larger for
the common case of GUI tools generating a usage message;
and similarly, make the default font used fixed-width else
we end up with garbled, difficult-to-read output.

Also, when multi-line, scrolled output is produced, scroll
back up to the start of the text, don't start with with an
offset at the end - the user will always want to read from
the start of the message.