Remove remaining non-signal-safe exit/abort calls in daemons
Frank spotted a few remaining calls to exit(3) on inappropriate
paths, audit once more and convert those to _exit and call any
safe cleanup routines by-hand (cannot rely on atexit in a signal
handler).  Similarly, some abort calls still remained in places
they should not.

Introduce a pattern of exiting with the signal number received
for the "unexpected" class of signal handlers to aid diagnosis.
Daemons return code is always zero or one today, hopefully this
will prove to be a generally useful diagnostic addition - it is
signal-safe too ;)