Additional services for automatic discovery
Add pmproxy and pmwebd to the set of services which are
advertised via Avahi.  Update pmfind to enable searching
on these, by default search for all known services.

Tests 727 and 950 are added to exercise that each daemon
is indeed advertising via avahi; common code from these
two tests and 840 (pmcd) is now in common.avahi.