Add a pmcd.services metric reflecting PCP daemon usage
After finding a need to quickly know which PCP services
are active on a host, on a number of occassions now, I've
finally cracked up and added a pmcd.services metric that
reflect this state.  A line that looks alot like:
 services: pmcd pmmgr pmproxy pmwebd
is added to pcp(1) output as well, using this metric when
it is available (else, output is as it was before).

This is implemented in a portable way by ensuring daemons
all manage pidfiles the same way pmcd does ($PCP_RUN_DIR),
so a helper function is added to libpcp for each server to
make use of.  While there, pulled the open-coded stackdump
functions from pmcd and pmie into a shared spot also since
this is more generally useful (they were inconsistent too,
with the pmie variant being relatively incomplete).

A small improvement is added to the common pidfile code in
that an atexit handler is now registered to cleanup, which
pmcd was lacking (ala the log footer writer).  In addition
the pidfiles are made readonly after initial creation.