Improvements in python timeval/metricspec error handling
Currently the python API has static methods on the pmContext
class allowing timeval and metricspec creation via the PMAPI
pmParseTimeInterval and pmParseMetricSpec interfaces.  We've
slightly botched the error handling however, and these calls
return a tuple with (result,errstr).  This is not a sensible
strategy, since the introduction of pmErr exceptions.

So, make the most of it and provide a better interface that
callers can use via new methods on the existing timeval and
MetricSpec classes.  The old, compat code simply calls the
new methods, and (continues to) returns an empty string for
the success case.  We also tack the error message onto the
exception for improved error handling, which was overlooked
last time through.

Existing tests exercise the compat case, and pmcollectl and
pmatop will be updated to use the new class methods.