Fix pmatop now in terms of sub-second recording
pmatop contains some duplicated code from pmcollectl (or
vice-versa) related to archive recording - also contains
the same bugs fixed recently in pmcollect wrt recording
for less than a second.  Make the code the same again, and
make a note to refactor this into the convenience classes
soon also.

This code, and pmcollectl, now make use of the new timeval
method for sleeping, and both are simpler as a result.

Update qa/722 to make use of very short sampling intervals
to exercise this change (and complete more quickly too).