Fix pmstat hostname reporting when running in cluster mode
Found a couple of bugs in the hostname reporting done by
pmstat and the pmGetContextHostName routine.  Firstly, in
pmstat a missing field initialisation in the context data
structure meant the hostname pointer could point off into
the weeds.  Secondly, the libpcp code that retrieves and
uses the pmcd.hostname metric didn't correctly setup the
context for pmLookupName/pmFetch with the result that the
original context was re-used and the wrong hostname ended
up being reported.

Add test qa/734 to exercise the whole horrible situation
and add some libpcp diagnostics around context switching
in pmGetContextHostName.