Extend python convenience classes for command line options
Focus here is on making the high level pmcc.MetricGroupManager
class able to take on more of the work that all client tools
have to do.  In particular, responding to command line options
like sample interval/count, end time, and so on impacts on the
logic needed around the fetch loop.  Add a MetricGroupPrinter
class which can be provided by a script, which will be called
upon to handle all matters related to output.

Ideally everything else will be handled by convenience classes
- this handles command line options and the fetch loop logic;
still to be further improved on (perhaps?) is the metric value
unpacking wrt counters and rate conversion.  Other helpers for
common output patters are also likely to be needed.