Improve handling of archives and errors in python pmcc module
Handling of errors in pmcc is much proved - previously it just
printed a message to stdout for all caught exceptions and then
continued right on its merry way (!).  This typically resulted
in promptly getting hit by a bus two or three lines on, due to
some assumptions about earlier success in setting up some data
structure.

Particularly problematic was archive replay reaching the end of
the archive, which is a natural and expected "error" condition.
Adding to the dilemma is the way we have good context for doing
error reporting at the initial exception catching site.  Handle
this by now printing to stderr, then raising SystemExit with an
exit code indicative of normal(EOL) or abnormal failure.  Then,
we catch this in the top level run() interface and return that
code to the script, which can deal with it as it will - usually
simply by exiting with that code, I expect.

Test qa/980 now passes.  Updated test 979 (live mode) to use a
shorter sampling interval while in that neck of the woods - to
complete more quickly but also exercise interval setting.