pmwebd: add CORS Access-Control-Allow-Origin: * to JSON response headers
This is necessary to allow webapps not served by pmwebd's fileserver
to have problem-free cross-origin data access to the /pmapi/* goodies.
(Despite the name, it does not function as any sort of access control,
in that the pmapi data is produced.  It only lets the web browsers
consume the data with less hassle.)

As a bonus, fix blinkenlights.js to be flexible with pmwebd port
numbering, which helps testing.  To check the CORS functionality,
one would hand-edit that file to hardcode some remote pmwebd server
and confirm that the web browser is willing & able to connect.