Add mem.util.available.
This is from /proc/meminfo in recent kernels (including RHEL7GA).
It the amount of memory that is available for a new workload, without
pushing the system into swap - estimated from MemFree, Active(file),
Inactive(file), and SReclaimable, as well as the "low" watermarks from
/proc/zoneinfo. See https://lkml.org/lkml/2013/11/7/264 for details.
On downrev kernels, returns No Values Available.

	modified:   src/pmdas/linux/help
	modified:   src/pmdas/linux/pmda.c
	modified:   src/pmdas/linux/proc_meminfo.c
	modified:   src/pmdas/linux/proc_meminfo.h
	modified:   src/pmdas/linux/root_linux