Add archive folio parsing mechanism into option handling code
After recent discussion with Stan about the needs of python tools
that read and write archive folios, this commit adds a mechanism
for automatically opening a folio and setting up struct pmOptions
with the archive(s) it contains.  While there, the exising host
and archive list mechanisms are tested further, and added to the
automatically available set (as long options only to not conflict
on existing short option characters).

Extensions are made to the pmgetopt shell utility to access this
functionality, and the python wrapper APIs.  A new qa/899 test is
tossed into the mix, and other existing tests are updated to make
sure we have decent functional coverage.  Also, a new QA group is
added to make it easier to run all the getopts tests at once.