Error handling and QA testing work for the nvidia PMDA
Updates to the PMDA itself include installation of the DSO
form, several bug fixes from my earlier round of updates and
the addition of error handling around all NVML calls.

Tests qa/744 and qa/745 have been added, as well as a simple
little shared library implementing the basic NVML interfaces
with a static configuration for 2 GPUs (for testing).  These
tests bail out if a system nvidia-ml.so is found, so that'd
be an interesting third case (which I cannot test here, so
I've not written the test).