libpcp/check-statics - expr(1) snarfoo fix up
Notices this little gem in Logs/pcp after a build on Centos 5.10

expr: warning: unportable BRE: `^.*\\.o$': using `^' as the first character
of the basic regular expression is not portable; it is being ignored

Strictly "^" is not needed for expr(1), although why warning about
it helps man or beast is beyond me.

Also added a guard for no object files while I was here.