pmdagfs2: Set default trace_pipe buffer size and allow control
Set a default value for the size of the trace_pipe buffer (per cpu) upon the
init of the PMDA. Have also created a control metric which allows the user to
set a custom size for buffer_size_kb up to 128MB (per cpu).

When there is high load on the GFS2 filesystems mounted there will be many
events captured in the trace_pipe, if the buffer is not sufficiently large
enough then entries may be overwritten before being read by the PMDA. By
increasing the size of the buffer we are minimizing the risk of this
occurring and ensuring we catch as many results as possible.