libapp - builds in main tree
No testing.
And lots of functional overlap between here and pmchart, so we'll
need so refactoring here to allow pmchart and pmview to share the
parts of libapp that are really commmon.
Also remnant references to pmchart here indicate origins of this
code on pmchart that were never cleaned up when it was extracted
to be used my pmview.