Ensure the same locale is set for sort and join commands
The join command checks the input is sorted using the same locale
as sort, differences here have been observed to cause join(1) to
fail with "join: file1 is not sorted" in test qa/128.  Audit all
of the joins and ensure they match their sorts.

Additionally, the perl scriptlet generating nfsstats was using an
incorrect string comparison operator, which was observed on still
other platforms to generate no reduce nfsstat output to nothing.