Abstract __pmConnect() into __pmConnect and __pmConnectWithFNDELAY().
In PCP, control over whether a connection attempt made via
__pmConnect() blocks or not is controlled by setting the
FNDELAY (O_NONBLOCK) file status flag in for the file
descriptor. However, NSPR's PR_Connect ignores the FNDELAY
flags (more accurately, always sets it) and uses an explicit
timeout argument to control whether the connection attempt
blocks and, if so, for how long. The NSPR implementation of
__pmConnect(), therefore has no way of knowing what the caller's
intent was.

To solve this, we need an explicit __pmConnectWithFNDELAY(). This
new function is internal to libpcp and not exported. It is called
from within __pmConnectTo() which is the function called within
PCP when non-blocking behavior is needed.

Code within PCP that calls __pmConnect() directly does not expect
non-blocking behavior and so does not need to handle EINPROGRESS.
This commit reverts a previous change
(commit commit a705d6bb86d87fa0cdf0fb0accf5299212cef08d)
for which such handling was added.

This commit fixes regressions in qa tests 230 and 835.