Attempt to workaround a GNU/Hurd build issue for Svante
Problem report is as follows, the pmieconf make rules are all
a bit wierd, with the subdir descent making things awkward.

From: Svante Signell <svante.signell@gmail.com>
"Due to some yet unknown reason the second time pmieconf is creating
the file 'local' it fails with the fname being empty in rules.c. The
attached patch makefile.patch is a work-around for this problem by not
deleting that file the second time.

(cd src/pmieconf; PCP_CONF=../../src/include/pcp.conf
LD_LIBRARY_PATH=../../src/libpcp/src:../../src/libpcp_pmda/src:$LD_LIBRARY_PATH
HOME=`pwd` PCP_ALT_CPP=../../src/pmcpp/pmcpp ./pmieconf -F -r rulesdir
-f local) manually works to create 'local', but not in the nested
make calls.

It seems that the re-creation of that file could be avoided by
modifying the files GNUmakefile and/or GNUmakefile.rules resulting in:
Nothing to be done for 'default_pcp' when getting to the examples
directory? (Dunno how to do that yet)."

For now, I've rearranged the targets a little to try to avoid
the problem and included parts of Svante's approach too, but the
dependencies here are still not quite right and several things
are always rebuilt when they shouldn't be (with or without this
change, that is).