Enforce properties of 'flags' parameter to __pmDiscoverServicesWithOptions().
This parameter is now declared: const volatile unsigned *flags

This enforces the requirement that __pmDiscoveryServicesWithOptions()
must not write the the memory referenced by this pointer, since
there is no knowledge of whether this memory requires synchronous
access protection.

It also enforces the semantics that the contents of the memory can
change at any time; specifically the 'interrupted' bit.