pmdadmcache: a device mapper cache target PMDA
Adds in the dmcache PMDA.  This exports metrics from the Linux
device mapper subsystem, when one or more cache targets are in
use.  These targets allow for faster cache-ier devices to sit
in front of slower, spinnier devices (e.g. SSD<>disk).  Metrics
about read/write cache hits/misses, various sizes of several of
the components, and so on, are exported by the kernel.

Tests qa/985 and qa/986 are added to exercise the new PMDA.  A
handful of updates are made to the python module code - mostly
issues found after running pylint on it once more.