pmdalinux: allow for mock testing with canned proc/sysfs files
As Mark is working toward testing device mapper stuff now, the
need for a test-data-injection mechanism for pmdalinux came up
once more - this commit introduces the first wave.  All of the
device code is converted over now, and some of the net and CPU
metrics.  Enough for Mark to get started though, and for me to
continue on in parallel.