pmwebd: extend precision for some floating point outputs
In particular, when emitting the PMWEBAPI json values for
pmExtractValue'd atoms, preserve the maximum number of precision
needed for their decimal representation, as kindly supplied by the
std::numeric_limits<TYPE>::digits10 template value.  Note that this is
not necessary in the graphite case as those numbers are
single-precision floats for whom the default works.