qa - ugly hack for qmc executables issue
Very ugly picker for executables from a build or a testsuite
installation.

Problem is that /var/lib/pcp/testsuite/qa is populated from the
packaging, but in a build context the qmc headers are defined
in ../../src/libpcp_qmc/src but the qmc library and the standard
PCP headers are not in ../../src they are out in the build area
c/o Makepkgs.

This pretty much makes it impossible to rebuild the qmc QA
executables from the source tree unless a full PCP make has
already been run there.

In QA common, we try the "make" recipe.  If this does not work,
fallback to the ugly hack in qt/setup-executables.