libpcp_gui - better cleanup after mkstemp() et al
For a long time, QA has been leaving empty files with mkstemp-like
names, e.g. tEvTkZ, in the current directory.

Finally tracked this down to the pmRecordSetup() method in libpcp_gui
that creates the temporary file, then uses it as the basename for
several files, but does not go back and unlink the (unused) temporary
file.