Delay the initialzation of NSPR, NSS and SASL libraries.
Initialization of these libraries incurs a significant
penalty in the form of resource (memory) usage. This change
delays this initialization, for both clients and servers,
until their respective funcionalities are required.

A significant amount of socket I/O abstraction code
(pre-connect/pre-accept NSPR implementation) was made obsolete
and was removed. The separation of native and NSPR file
descriptors in the IPC table is also no longer necessary
and the related code has been removed.