qa/common.qt - make Qt dependent QA conditional
For many of the older platform and non-core distros, we cannot
build the GUI bits because Qt is not installed or is too old.

The build handles this OK.

QA did not.  This change uses the absence of an installed
libpcp_qmc.so as the guard to "notrun" any of the GUI related
QA tests.