Initial support for native systemd service files
Allow for the use of native systemd service files for the
pcp daemons (pmcd, pmproxy, pmmgr, pmwebd, pmlogger, pmie).
A --with[-out]-systemd configure option is added, and we're
default to "on" for systems where pkg-config finds the unit
dir (pkg-config systemd --variable=systemdsystemunitdir).

The approach used is to piggyback on the existing rc scripts
and use the "oneshot" systemd Service Type (a model borrowed
from stap-server in systemtap).  In the case where systemd is
being enabled the definition of PCP_RC_DIR is auto-configured
to the home of these scripts (which is no longer below /etc),
and they instead live alongside a library rc script that they
source.  This has the property of minimising the QA fallout,
wrt tests that use these scripts directly.

Use of systemctl is added into a number of places where we've
got dependence on chkconfig enabling/disabling/state-checking
also (esp. common.check in QA and rc-proc.sh).

This resolves Red Hat BZ #1044682 and Fedora BZ #996438.