pmwtf: add some new reporting options, man page updates and a test
Updates to the pmwtf tool to provide finer control over the report
(--skip-missing and --skip-excluded options) to suppress parts of
the output that may be uninteresting or overly verbose.

Install it into /usr/bin now too, keeping a compat symlink back in
PCP_BINADM_DIR for any scripts using it.