Add auto-deflate support for the xz compression format
Extend the existing libpcp support for auto-deflation of
compressed archives with xz(1).  This format seems to be
popular in oss projects (kernel.org, debian.org) as the
compression format of choice.

Go with the flow - this makes it the default compression
of choice for PCP too.  In addition, the crontabs that we
install default to compressing after 3 days; we neglected
to enable compression at all in the crontabs previously
(i.e. it was opt-in, but now its opt-out or opt-other for
any other compression format/tools you like).

A couple more compression file suffixes are added to the
set recognised in the daily scripts as well and the regex
to identify these is made a little more readable.

QA tests 435, 436 and 437 are refactored to share common
code relating to compression testing, and new test qa/958
is added to exercise xz(1) use.