Merge the pmgadgets WIP branch and begin refactoring code
The pmgadgets and pmview branches had both formed new libs
for sharing common code - merge these into one and prepare
the way for pmchart to play in this sandpit too.

At this stage alot of this is not built as part of regular
builds.  One visible change is that images being pulled in
via Qt resource files (qrc) are no longer individually set
in the makefiles.  Instead, each separate tool creates a
local symlink to the top-level images SUBDIR and they are
always referenced with an images/ prefix.  Simplifies some
parts of the build.