pmdagfs2: Fix issue with RHEL 6.6 ftrace not having irq-info option
The current version of ftrace included in the RHEL 6.6 beta does not include
the irq-info option as part of its trace function output (default is include
and enabled with other distributions). This results in the tracepoint output
given by the trace_pipe being different to what is expected and the PMDA
hanging on collection of the newly added latency metrics (timeout and killed
by PMCD).

To tackle this we turn off the irq-info option in versions of ftrace that
have the option and tweak the expected trace_pipe output so that the same
output will be given by all distributions if they have irq-info option
support or not.