Implement new PDU_FLAG_SECURE_ACK protocol for clients and servers.
When the server supports it, clients can request that an error
pdu indicacting the server's NSS and SSL initialization status
be sent by the server as part of the secure connection handshake.
This indicates whether the server is able to upgrade its connection
to SSL/TLS and, thus, whether the client should do the same.

Older servers will not indicate this capability during the initial
handshake and so the client will not request or expect the pdu.
Older clients will not request this and the server will therefore
not send the pdu.