qa/admin/allow-pmlc-access - QA Helper
For distributed QA to work, we need each QA machine to be able to
run pmlc to connect to the primary pmlogger on each other QA machine.

This script amends the config.default pmlogger configuration file to
allow unlimited pmlc access from hosts on the same IPv4 network.

Works for me in Carnegie ... others may need to extend to suit their
local needs.