qmake fix - ugly
qmake appears to be broken in that the generated Makefile contains a
bogus -L/some/lib/dir component _before_ the PCP and related lib
dirs we care about.

If a down-rev libpcp, for example, is installed on /some/lib/dir, we
try to link against the down-rev libpcp instead of the correct one in
the build ... discovered by Jeff Hanson @ SGI.

The fix is to edit the Makefile after qmake generates it and before we
use it in the build.