Older python support
Changes to accommodate different "except" clause syntax in Python
before version 2.5.

Source files containing "except .. as ..." are renamed from *.py
to *.pyin and a buildrule generates appropriate syntax in the *.py
files which are used in the packaging and then discarded.

For Python before 2.5, the except clause is rewritten as
"except .., ..." (comma replaces " as").