timeval refactoring - part 2, PMDAs and PMAPI clients
This group of commits refactors the way we handle struct timeval
and in particular conversion to and from a double.

Use the libpcp routines in preference to inline code.

The end result is we avoid arithmetic involving 1000000 appearing
all over the code base, and encapsulate the correct arithmetic in
just one place (tv.c of libpcp).