timeval refactoring - part 5, QA output
This group of commits refactors the way we handle struct timeval
and in particular conversion to and from a double.

Use the libpcp routines in preference to inline code.

After all the changes, we need to amend our view of what is the
"correct" output for a couple of tests ... the previous .out files
were just wrong, and this had not been noticed.