qa/common.check - tighter checks for pmlogger termination
In _wait_pmlogger_end() we were waiting for the control file to be
removed.  But even after cleaning up the stdio buffer flushing in
pmlogger there is still a race, because the trailer "finished"
message is written from libpcp routines in an atexit() handler which
is after the control file is removed.

Extend the test to wait for the pmlogger process to have really
exited, using "kill -s 0 $pid" as the test.