systemd pmda: use -Uroot if user adm not available
On some systems, it was reported that user "adm" is not installed, and
so systemd-pmda's default -Uadm behavior breaks it.  This new Install
script checks whether user adm exists, and if not, sticks with root.
This may reduce security somewhat (in case the pmda is broken into),
but the previous per-client uid/gid checks are believed effective.

This was hand-tested on a Fedora machine, with and without commenting
out the adm: account in /etc/passwd.  This change is probably too
invasive to make mechanically as a part of a qa test case.