telnet-probe: fix byte-by-byte copying
During the transition to the NSS libraries, a __pmWrite call replaced
a fputc when transcribing bytes to/from the remote socket.  This was
broken because the parameter to __pmWrite was an int rather than a
char, and so instead of sending 'a', telnet-probe sends 'a\0\0\0'.
Similarly on the socket-receive side, we tried to read ints rather than
chars via __pmRead.  We now narrow the int-carrying values to chars
where they belong.