pmwebapi: ACAO and QA improvements
It was reported that pmwebd should send a Access-Control-Allow-Origin: *
response header even for error codes, so that browsers permit web apps
to display them.  While hand-testing this, ran into a problem where an
assertion-failure was triggerable from a bad HTTP request.  Fixed that.

On the QA side, we enhance the 660 test case:
- running curl with -v a few times, to get the ACAO:* on error responses
- running telnet-probe to generate some invalid http requests