man pages: move build-conditional man pages alongside sources
Currently we duplicating the logic for choosing which man pages
get included in the build in two places - this continues the
trend of removing that dup'd makefile logic by keeping sources
and man page within one makefile for those situations.

Soon this will allow the man1 makefile to not explictly list
each and every man page as well, not quite there yet though.