man pages: move Linux-specific perl pod to static man pages
We are currently missing some man pages from the set that is
now being indexed and served online.  One reason is that the
text is (in some cases) hidden away in perl scripts.  Theres
not alot of value being gained in doing that, so I've begun
moving those into separate files such that the online indexing
can easily find them (without the need to perform, a build).

This commit handles just the Linux-specific perl scripts.