pmdapapi: move the helptext buffer off the stack (fixes qa/914)
Move the buffer that holds the helptext path name off the stack.
This was causing problems under higher gcc optimisation levels,
when the pmdaInterface e_helptext pointer ended up spontaneously
pointing into weeds on the call to papi_internal_init().

This resolves failure of qa/914 when pmdapapi is built -O2, as
it otherwise contains pminfo/pmstore warnings of the form...
[DATE] pminfo(PID) Warning: pmdaInit: PMDA papi DSO: Unable to open help text file(s) from "": No such file or directory