qa - run 994 to check-n-repair permissions and ownership after each test
Following Nathan's excellent suggestion, check now runs 994 after
each test to check that key files and directories retain the ownership
and mode as at install time.  994 now also supports a --fix option that
repairs any breakage, and check uses this.

If a test breaks the permissions or modes, it is considered a failure
and the details are appended to the $seq.out.bad file.