debian packaging - rework pmie and pmlogger control and config files
The control files for pmie and pmlogger (used by the *_daily
and *_check scripts to determine what pmie and pmlogger
instances to start or restart) stay in $PCP_SYSCONF_DIR/pmie and
$PCP_SYSCONF_DIR/pmlogger, but these directories and the control
files are owned by root.

The search directory order for pmie (or pmlogger) config files
(from the -c option) that are not absolute pathnames is "." then
$PCP_VAR_DIR/config/pmie (or pmlogger).  These directories are
owned and writeable by pcp, so pmlogger_check can create/recreate
the config.default file using pmlogconf.

This commit contains the postinstall changes to enforce the new
rules and ease migration for upgrades.