qa - heaps - config file changes
Do a better job of saving and restoring configuration files (or
directories) in a way that (a) avoid mv across mount points, (b)
preserves the mode and ownership of the files, and (c) leaves a
named body in the directory containing the config file or directory
if the QA test aborts (so we know who to blame).

Use _save_config <pathname> to move <pathname> to <pathname>.$seq
and make <pathname> a copy that can be munged in QA.

Use _restore_config <pathname> to reverse the operation.