Make adding metrics to the current eventset more robust
If we attempt to add an event that cannot be added (for example,
if two metrics can't be counted at the same time), we should make
the attempt, but still restart papi and run the other metrics if
possible.  If debugging is turned on, output the error message to
the log file.