libpcp/AF.c - fix lock initialization race
Sometimes seen in qa/134 (and others), a pmlogger -c /dev/null -L would
hang and not allow pmlc to connect.

Tracked down to a nested __pmInitLocks() call because the timer alarm
went off in the middle of setting up.  This fix forces __pmInitLocks()
to be called early on when the __pmAF* routines are called.