Makepkgs: on RH platforms, use rpm --eval ... for fuller ./configure
The basic set of configure defaults in Makepkgs doesn't account for
platform specifics, like where 64-bit libraries go.  This can result
in packages that install parts in directories that mismatch the
platform rules, and can cause accidental conflicts later.  So, to
future-proof this part, on RH platforms we transcribe a portion of
the %configure macro into Makepkgs, so e.g. --libdir and many other
friends are set.  It is likely that other RPM-based platforms could
use a similar hack to make Makepkgs base-os-compatible.

Tested on F19 x86-64 and RHEL5 i686.