qa/260 - refactor choice of metrics
We were previously using pmcd.pdu_in.* and pmcd.pdu_out.* for the
derived expressions and the pmie expressions.  This does not work in
the presence of concurrent access to pmcd outside this QA test, e.g.
a local system pmie or pmlogger or some bozo from a remote machine.

So, change to sample.const_rate.* and sampledso.const_rate so we can
control the rate of counter increase and the values of the counters
in ways that make the test more likely to pass.