qa: fix test qa/660 so it can pass when run from qa/check
Fix a bug in the exit status handling in test 660 that prevents it
from ever having passed when run from the testsuite harness.  The
cleanup routine was being called before setting the exit status to
zero, so we were never returning with a pass status.

Comment out the telnet-probe invocations, they hang in read(2) for
me on rhel6 (will split these into a separate test shortly).  Also
in preparation for test splitting, refactor some common code into
the common.webapi script.

Update filters so that the order of IPv4/IPv6 address resolution
is no longer significant, and test output is more deterministic.