qa/518: stretch timing to run more reliably
This test case involves a race between pmie running rules a certain
number of times, and a separate task waiting (by time) to match the
results via /usr/bin/pcp.  It was observed as flakey, not always
winning the race.  This version stretches the pmie time and fine-tunes
the waiting interval to give the race a larger (multi-second) window
for success.