papi pmda: add counter auto-enable on fetch
By reinterpreting the papi_info[].metric_enabled as a timestamp rather
than a boolean, and a small bit of logic, we get automatic enablement
of PAPI metric/counters upon pmfetch.  This allows PAPI metrics to be
used with plain old pmlogger, pmval, etc., without necessary manual
pmstore's.  Counters are retired based on a timeout mechanism,
assisted by pmaf(3).  The papi.control.status metric is enhanced to
show the remaining lifespan of the counters, and to make it more safe
& reliable.

A new papi.control.auto_enable metric is available to set the timeout
(default 120 seconds).  The Install script runs a papi.control.reset
operation to avoid leaving the counters running for even that long
after the post-install metric/value checks.

The previous pmstore mechanism is left in place unchanged, and
overrides the automatic scheme.